requirements.txt
google-cloud-aiplatform>=1.38.0
rich>=13.7.0
pyyaml>=6.0.1